runtime/noos: Working tickless scheduler.

+++ egroot/src/time/time-noos.go
@@ -8,8 +8,8 @@ var start Time
 
 func Set(t Time) {
 	up := rtos.Uptime()
-	sec := int64(up / 1e9)
-	nsec := int32(up - uint64(sec)*1e9)
+	sec := up / 1e9
+	nsec := int32(up - sec*1e9)
 	if nsec < 0 {
 		sec--
 		nsec += 1e9
@@ -19,8 +19,8 @@ func Set(t Time) {
 
 func now() (sec int64, nsec int32) {
 	up := rtos.Uptime()
-	sec = int64(up / 1e9)
-	nsec = int32(up - uint64(sec)*1e9)
+	sec = up / 1e9
+	nsec = int32(up - sec*1e9)
 	sec = start.sec + sec
 	nsec = start.nsec + nsec
 	if nsec >= 1e9 {
